Split query commit runner helpers
