+++ crates/query/src/runner/commits.rs
@@ -7,529 +7,35 @@
 //! - `generate_commit_group_key()` / `json_value_to_key()` / `build_commits_mapping()`
 //! - `commit_to_fields()` / `compare_json_values()`
 
-use document::Document;
 use identity::Did;
 use serde_json::Value as JsonValue;
 use std::collections::HashSet;
 
 use crate::error::Result;
-use crate::mapper::{Aggregate, AggregateType, Requestable, Select};
+use crate::mapper::{Requestable, Select};
 use crate::txn::TransactionRegistry;
 
 use super::commits_height::{extract_commits_height_range, HeightRangeExtraction};
+use super::{DocFetcher, QueryRunner};
+
+#[cfg(test)]
 use super::commits_numeric::{
     max_commit_numeric_values, min_commit_numeric_values, sum_commit_numeric_values,
     CommitNumericValue,
 };
-use super::{DocFetcher, QueryRunner};
 
-fn is_commit_aggregate_only_selection(select: &Select) -> bool {
-    !select.fields.is_empty()
-        && select
-            .fields
-            .iter()
-            .all(|field| matches!(field, Requestable::Aggregate(_)))
-}
+mod aggregate;
+mod grouping;
+mod nested_filter;
+mod rendering;
+
+use aggregate::is_commit_aggregate_only_selection;
+
+// Every formerly private helper moved into a child uses `pub(super)`. This
+// preserves its original effective visibility to the `commits` module and all
+// descendants without exposing it outside that module tree.
 
 impl<F: DocFetcher + 'static, R: TransactionRegistry> QueryRunner<F, R> {
-    fn commit_aggregate_target_field<'a>(&self, agg: &'a Aggregate) -> Option<&'a str> {
-        agg.targets.first().and_then(|target| {
-            target
-                .field_name
-                .as_deref()
-                .filter(|field_name| !field_name.is_empty())
-                .or_else(|| {
-                    Some(target.host_name.as_str()).filter(|host_name| !host_name.is_empty())
-                })
-        })
-    }
-
-    fn normal_value_to_commit_numeric(
-        &self,
-        value: &document::NormalValue,
-    ) -> Option<CommitNumericValue> {
-        value.as_int().map(CommitNumericValue::Int).or_else(|| {
-            value
-                .as_float64()
-                .map(CommitNumericValue::Float)
-                .or_else(|| {
-                    value
-                        .as_float32()
-                        .map(|value| CommitNumericValue::Float(value as f64))
-                })
-        })
-    }
-
-    fn decode_commit_delta_numeric(&self, commit: &Document) -> Option<CommitNumericValue> {
-        use base64::Engine;
-
-        let delta_base64 = commit.get("delta")?.as_str()?;
-        let bytes = base64::engine::general_purpose::STANDARD
-            .decode(delta_base64)
-            .ok()?;
-        let value = ciborium::from_reader::<document::NormalValue, _>(&bytes[..]).ok()?;
-        self.normal_value_to_commit_numeric(&value)
-    }
-
-    fn commit_numeric_value(
-        &self,
-        commit: &Document,
-        field_name: &str,
-    ) -> Option<CommitNumericValue> {
-        if field_name == "delta" {
-            self.decode_commit_delta_numeric(commit)
-        } else {
-            commit
-                .get(field_name)
-                .and_then(|value| self.normal_value_to_commit_numeric(value))
-        }
-    }
-
-    fn collect_commit_numeric_values(
-        &self,
-        commit: &Document,
-        group_docs: Option<&[Document]>,
-        field_name: &str,
-    ) -> Vec<CommitNumericValue> {
-        let mut values = Vec::new();
-
-        if let Some(group_docs) = group_docs {
-            for doc in group_docs {
-                if let Some(value) = self.commit_numeric_value(doc, field_name) {
-                    values.push(value);
-                }
-            }
-        } else if let Some(value) = self.commit_numeric_value(commit, field_name) {
-            values.push(value);
-        }
-
-        values
-    }
-
-    fn count_commit_aggregate_values(
-        &self,
-        commit: &Document,
-        group_docs: Option<&[Document]>,
-        field_name: Option<&str>,
-    ) -> i64 {
-        match field_name {
-            None => group_docs.map(|docs| docs.len()).unwrap_or(1) as i64,
-            Some(field_name) => {
-                let mut count = 0i64;
-
-                let mut count_doc = |doc: &Document| {
-                    if field_name == "delta" {
-                        if self.decode_commit_delta_numeric(doc).is_some() {
-                            count += 1;
-                        }
-                        return;
-                    }
-
-                    if let Some(value) = doc.get(field_name) {
-                        if let Ok(json_value) = crate::json_convert::normal_value_to_json(value) {
-                            if let Some(array) = json_value.as_array() {
-                                count += array.len() as i64;
-                            } else if !json_value.is_null() {
-                                count += 1;
-                            }
-                        }
-                    }
-                };
-
-                if let Some(group_docs) = group_docs {
-                    for doc in group_docs {
-                        count_doc(doc);
-                    }
-                } else {
-                    count_doc(commit);
-                }
-
-                count
-            }
-        }
-    }
-
-    fn compute_commit_aggregate(
-        &self,
-        agg: &Aggregate,
-        commit: &Document,
-        group_docs: Option<&[Document]>,
-    ) -> JsonValue {
-        match agg.aggregate_type {
-            AggregateType::Count => {
-                let count = self.count_commit_aggregate_values(
-                    commit,
-                    group_docs,
-                    self.commit_aggregate_target_field(agg),
-                );
-                JsonValue::Number(count.into())
-            }
-            AggregateType::Sum => {
-                let Some(field_name) = self.commit_aggregate_target_field(agg) else {
-                    return JsonValue::Number(0.into());
-                };
-                let values = self.collect_commit_numeric_values(commit, group_docs, field_name);
-                sum_commit_numeric_values(&values)
-            }
-            AggregateType::Average => {
-                let Some(field_name) = self.commit_aggregate_target_field(agg) else {
-                    return JsonValue::Number(0.into());
-                };
-                let values = self.collect_commit_numeric_values(commit, group_docs, field_name);
-                let avg = if values.is_empty() {
-                    0.0
-                } else {
-                    values.iter().map(|value| value.as_f64()).sum::<f64>() / values.len() as f64
-                };
-                serde_json::Number::from_f64(avg)
-                    .map(JsonValue::Number)
-                    .unwrap_or_else(|| JsonValue::Number(0.into()))
-            }
-            AggregateType::Min => {
-                let Some(field_name) = self.commit_aggregate_target_field(agg) else {
-                    return JsonValue::Null;
-                };
-                let values = self.collect_commit_numeric_values(commit, group_docs, field_name);
-                min_commit_numeric_values(&values)
-            }
-            AggregateType::Max => {
-                let Some(field_name) = self.commit_aggregate_target_field(agg) else {
-                    return JsonValue::Null;
-                };
-                let values = self.collect_commit_numeric_values(commit, group_docs, field_name);
-                max_commit_numeric_values(&values)
-            }
-        }
-    }
-
-    /// Render a Document's fields as a JSON object using only the fields requested by a Select.
-    pub(crate) fn render_document_fields(
-        &self,
-        doc: &Document,
-        select: &Select,
-    ) -> serde_json::Map<String, JsonValue> {
-        let mut obj = serde_json::Map::new();
-        for field in &select.fields {
-            if let Requestable::Field(f) = field {
-                let fname = &f.name;
-                let output = f.output_name();
-                if fname == "_docID" {
-                    if let Some(id) = doc.id() {
-                        obj.insert(output.to_string(), JsonValue::String(id.to_string()));
-                    } else {
-                        obj.insert(output.to_string(), JsonValue::Null);
-                    }
-                } else if fname == "__typename" {
-                    obj.insert(
-                        output.to_string(),
-                        JsonValue::String(select.collection_name.clone()),
-                    );
-                } else if let Some(nv) = doc.get(fname) {
-                    let json_val =
-                        crate::json_convert::normal_value_to_json(nv).unwrap_or(JsonValue::Null);
-                    obj.insert(output.to_string(), json_val);
-                } else {
-                    obj.insert(output.to_string(), JsonValue::Null);
-                }
-            }
-        }
-        obj
-    }
-
-    /// Fetch version (commit) data for a document.
-    ///
-    /// Returns an array of commit objects filtered to composite commits (fieldName = "_C")
-    /// and rendered with the requested fields from the _version selection.
-    pub(crate) async fn fetch_version_data(
-        &self,
-        fetcher: &dyn DocFetcher,
-        doc_id: &str,
-        version_select: &Select,
-        collection_id: &str,
-        target_cid: Option<&str>,
-    ) -> Result<JsonValue> {
-        use crate::fetcher::CommitsQueryOptions;
-
-        // For CID queries, traverse deeply from the specific CID block.
-        // For all other cases (mutation results, regular queries), None means
-        // unlimited DAG traversal from heads - returning all versions.
-        let depth = if target_cid.is_some() {
-            Some(1000)
-        } else {
-            None
-        };
-
-        let options = CommitsQueryOptions {
-            doc_id: Some(doc_id.to_string()),
-            cid: target_cid.map(|s| s.to_string()),
-            depth,
-            height_start: None,
-            height_end: None,
-            field_name: None,
-        };
-
-        let commits = fetcher.get_commits(&options).await?;
-
-        // Filter to composite commits only (fieldName = "_C")
-        // and render the requested fields
-        let mut version_results: Vec<JsonValue> = Vec::new();
-
-        for mut commit in commits {
-            // Filter to composite commits
-            let field_name = commit
-                .get("fieldName")
-                .and_then(|v| v.as_str())
-                .unwrap_or("");
-            if field_name != "_C" {
-                continue;
-            }
-
-            commit.set("collectionID", collection_id.to_string());
-            let commit_json = self.render_commit(&commit, version_select)?;
-            version_results.push(commit_json);
-        }
-
-        // Sort by height descending (newest first)
-        version_results.sort_by(|a, b| {
-            let h_a = a.get("height").and_then(|v| v.as_i64()).unwrap_or(0);
-            let h_b = b.get("height").and_then(|v| v.as_i64()).unwrap_or(0);
-            h_b.cmp(&h_a)
-        });
-
-        Ok(JsonValue::Array(version_results))
-    }
-
-    /// Render a commit document according to the _version selection fields.
-    fn render_commit(&self, commit: &Document, version_select: &Select) -> Result<JsonValue> {
-        let mut obj = serde_json::Map::new();
-
-        for requestable in &version_select.fields {
-            match requestable {
-                Requestable::Field(f) => {
-                    let field_name = &f.name;
-                    let output_name = f.output_name();
-
-                    if let Some(value) = commit.get(field_name) {
-                        let json_value = crate::json_convert::normal_value_to_json(value)
-                            .unwrap_or(JsonValue::Null);
-                        obj.insert(output_name.to_string(), json_value);
-                    } else {
-                        obj.insert(output_name.to_string(), JsonValue::Null);
-                    }
-                }
-                Requestable::Select(nested) => {
-                    let field_name = &nested.field.name;
-                    let output_name = nested.field.output_name();
-
-                    // Handle nested selections (links, heads) with optional filter
-                    if let Some(value) = commit.get(field_name) {
-                        if let Ok(json_val) = crate::json_convert::normal_value_to_json(value) {
-                            if let Some(arr) = json_val.as_array() {
-                                // Apply filter if present on the nested selection
-                                let filtered_items: Vec<&JsonValue> =
-                                    if let Some(ref filter) = nested.filter {
-                                        arr.iter()
-                                            .filter(|item| {
-                                                // Check each filter condition against the item
-                                                self.json_item_matches_filter(item, filter)
-                                            })
-                                            .collect()
-                                    } else {
-                                        arr.iter().collect()
-                                    };
-
-                                let nested_results: Vec<JsonValue> = filtered_items
-                                    .into_iter()
-                                    .map(|item| {
-                                        let mut nested_obj = serde_json::Map::new();
-                                        for nested_field in &nested.fields {
-                                            if let Requestable::Field(nf) = nested_field {
-                                                let nf_name = &nf.name;
-                                                let nf_output = nf.output_name();
-                                                if let Some(nv) = item.get(nf_name) {
-                                                    nested_obj
-                                                        .insert(nf_output.to_string(), nv.clone());
-                                                } else {
-                                                    nested_obj.insert(
-                                                        nf_output.to_string(),
-                                                        JsonValue::Null,
-                                                    );
-                                                }
-                                            }
-                                        }
-                                        JsonValue::Object(nested_obj)
-                                    })
-                                    .collect();
-                                obj.insert(
-                                    output_name.to_string(),
-                                    JsonValue::Array(nested_results),
-                                );
-                            } else {
-                                obj.insert(output_name.to_string(), JsonValue::Null);
-                            }
-                        } else {
-                            obj.insert(output_name.to_string(), JsonValue::Null);
-                        }
-                    } else {
-                        obj.insert(output_name.to_string(), JsonValue::Array(vec![]));
-                    }
-                }
-                Requestable::Aggregate(agg) => {
-                    let output_name = agg.output_name();
-                    obj.insert(
-                        output_name.to_string(),
-                        self.compute_commit_aggregate(agg, commit, None),
-                    );
-                }
-                Requestable::Similarity(_) => {
-                    // Similarity is not applicable in commit context
-                }
-                Requestable::FullTextSearch(_) => {
-                    // Full-text search is not applicable in commit context
-                }
-            }
-        }
-
-        Ok(JsonValue::Object(obj))
-    }
-
-    /// Check if a JSON object matches a filter for nested commit selections.
-    ///
-    /// This is a simplified filter matcher for nested selections like `links(filter: {fieldName: {_eq: "Age"}})`.
-    /// The filter conditions are stored as `{field_name: {_op: value}}`.
-    fn json_item_matches_filter(&self, item: &JsonValue, filter: &crate::mapper::Filter) -> bool {
-        use crate::mapper::FilterOp;
-
-        // Get the filter conditions - a map of field_name -> operator conditions
-        let conditions = filter.conditions();
-
-        for (field_name, condition_value) in conditions {
-            // Check if this is a logical operator (_and, _or, _not)
-            if let Some(op) = FilterOp::parse(field_name) {
-                match op {
-                    FilterOp::And => {
-                        if let JsonValue::Array(arr) = condition_value {
-                            for sub_cond in arr {
-                                if let JsonValue::Object(obj) = sub_cond {
-                                    let sub_filter =
-                                        crate::mapper::Filter::from_conditions(obj.clone());
-                                    if !self.json_item_matches_filter(item, &sub_filter) {
-                                        return false;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    FilterOp::Or => {
-                        if let JsonValue::Array(arr) = condition_value {
-                            let mut any_match = false;
-                            for sub_cond in arr {
-                                if let JsonValue::Object(obj) = sub_cond {
-                                    let sub_filter =
-                                        crate::mapper::Filter::from_conditions(obj.clone());
-                                    if self.json_item_matches_filter(item, &sub_filter) {
-                                        any_match = true;
-                                        break;
-                                    }
-                                }
-                            }
-                            if !any_match {
-                                return false;
-                            }
-                        }
-                    }
-                    FilterOp::Not => {
-                        if let JsonValue::Object(obj) = condition_value {
-                            let sub_filter = crate::mapper::Filter::from_conditions(obj.clone());
-                            if self.json_item_matches_filter(item, &sub_filter) {
-                                return false;
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-                continue;
-            }
-
-            // This is a field condition: field_name -> {_op: value}
-            let item_value = item.get(field_name);
-
-            // The condition_value should be an object like {"_eq": "Age"}
-            if let JsonValue::Object(ops) = condition_value {
-                for (op_name, expected_value) in ops {
-                    if let Some(op) = FilterOp::parse(op_name) {
-                        let matches = self.check_filter_op(item_value, op, expected_value);
-                        if !matches {
-                            return false;
-                        }
-                    }
-                }
-            }
-        }
-
-        true
-    }
-
-    /// Check if an item value matches a filter operator condition.
-    fn check_filter_op(
-        &self,
-        item_value: Option<&JsonValue>,
-        op: crate::mapper::FilterOp,
-        expected: &JsonValue,
-    ) -> bool {
-        use crate::mapper::FilterOp;
-
-        match op {
-            FilterOp::Eq => match (item_value, expected) {
-                (Some(JsonValue::String(a)), JsonValue::String(b)) => a == b,
-                (Some(JsonValue::Number(a)), JsonValue::Number(b)) => a == b,
-                (Some(JsonValue::Bool(a)), JsonValue::Bool(b)) => a == b,
-                (Some(JsonValue::Null), JsonValue::Null) => true,
-                (None, JsonValue::Null) => true,
-                _ => false,
-            },
-            FilterOp::Ne => match (item_value, expected) {
-                (Some(JsonValue::String(a)), JsonValue::String(b)) => a != b,
-                (Some(JsonValue::Number(a)), JsonValue::Number(b)) => a != b,
-                (Some(JsonValue::Bool(a)), JsonValue::Bool(b)) => a != b,
-                (Some(JsonValue::Null), JsonValue::Null) => false,
-                (None, _) => true,
-                _ => true,
-            },
-            FilterOp::Gt => match (item_value.and_then(|v| v.as_f64()), expected.as_f64()) {
-                (Some(a), Some(b)) => a > b,
-                _ => false,
-            },
-            FilterOp::Gte => match (item_value.and_then(|v| v.as_f64()), expected.as_f64()) {
-                (Some(a), Some(b)) => a >= b,
-                _ => false,
-            },
-            FilterOp::Lt => match (item_value.and_then(|v| v.as_f64()), expected.as_f64()) {
-                (Some(a), Some(b)) => a < b,
-                _ => false,
-            },
-            FilterOp::Lte => match (item_value.and_then(|v| v.as_f64()), expected.as_f64()) {
-                (Some(a), Some(b)) => a <= b,
-                _ => false,
-            },
-            FilterOp::In => {
-                if let JsonValue::Array(values) = expected {
-                    item_value.map(|v| values.contains(v)).unwrap_or(false)
-                } else {
-                    false
-                }
-            }
-            FilterOp::Nin => {
-                if let JsonValue::Array(values) = expected {
-                    item_value.map(|v| !values.contains(v)).unwrap_or(true)
-                } else {
-                    true
-                }
-            }
-            _ => true, // For unsupported operators, default to matching
-        }
-    }
-
     /// Execute a _commits system collection query.
     ///
     /// This handles queries to the special _commits collection which fetches
@@ -1044,131 +550,6 @@ impl<F: DocFetcher + 'static, R: TransactionRegistry> QueryRunner<F, R> {
         }
 
         Ok(JsonValue::Array(results))
-    }
-
-    /// Generate a group key from commit field values.
-    /// Format matches Go DefraDB: `{index}_{value}_` for each field.
-    fn generate_commit_group_key(commit: &document::Document, fields: &[String]) -> String {
-        let mut key = String::new();
-        for (i, field_name) in fields.iter().enumerate() {
-            key.push_str(&format!("{}_", i));
-            if let Some(value) = commit.get(field_name) {
-                if let Ok(json_val) = crate::json_convert::normal_value_to_json(value) {
-                    key.push_str(&Self::json_value_to_key(&json_val));
-                } else {
-                    key.push_str("null");
-                }
-            } else {
-                key.push_str("null");
-            }
-            key.push('_');
-        }
-        key
-    }
-
-    /// Convert a JSON value to a string for use in group key.
-    fn json_value_to_key(value: &JsonValue) -> String {
-        match value {
-            JsonValue::Null => "null".to_string(),
-            JsonValue::Bool(b) => b.to_string(),
-            JsonValue::Number(n) => n.to_string(),
-            JsonValue::String(s) => s.clone(),
-            JsonValue::Array(arr) => {
-                format!(
-                    "[{}]",
-                    arr.iter()
-                        .map(Self::json_value_to_key)
-                        .collect::<Vec<_>>()
-                        .join(",")
-                )
-            }
-            JsonValue::Object(obj) => {
-                format!(
-                    "{{{}}}",
-                    obj.iter()
-                        .map(|(k, v)| format!("{}:{}", k, Self::json_value_to_key(v)))
-                        .collect::<Vec<_>>()
-                        .join(",")
-                )
-            }
-        }
-    }
-
-    /// Check if a JSON value matches a filter.
-    /// Used for filtering nested arrays like links and heads in commit queries.
-    fn matches_json_filter(value: &JsonValue, filter: &crate::Filter) -> bool {
-        filter.matches_json_object(value).unwrap_or(false)
-    }
-
-    /// Build a DocumentMapping for commit fields.
-    fn build_commits_mapping() -> crate::document::DocumentMapping {
-        let mut mapping = crate::document::DocumentMapping::new();
-        mapping.add(0, "cid");
-        mapping.add(1, "height");
-        mapping.add(2, "fieldName");
-        mapping.add(3, "docID");
-        mapping.add(4, "delta");
-        mapping.add(5, "collectionVersionId");
-        mapping.add(6, "links");
-        mapping.add(7, "heads");
-        mapping.add(8, "signature");
-        mapping
-    }
-
-    /// Convert a commit document to a fields array for filter evaluation.
-    fn commit_to_fields(
-        commit: &document::Document,
-        _mapping: &crate::document::DocumentMapping,
-    ) -> Vec<Option<JsonValue>> {
-        let field_names = [
-            "cid",
-            "height",
-            "fieldName",
-            "docID",
-            "delta",
-            "collectionVersionId",
-            "links",
-            "heads",
-            "signature",
-        ];
-        field_names
-            .iter()
-            .map(|name| {
-                commit
-                    .get(name)
-                    .and_then(|v| crate::json_convert::normal_value_to_json(v).ok())
-            })
-            .collect()
-    }
-
-    /// Compare two JSON values for ordering.
-    fn compare_json_values(
-        a: Option<&document::NormalValue>,
-        b: Option<&document::NormalValue>,
-    ) -> std::cmp::Ordering {
-        use std::cmp::Ordering;
-
-        match (a, b) {
-            (None, None) => Ordering::Equal,
-            (None, Some(_)) => Ordering::Less,
-            (Some(_), None) => Ordering::Greater,
-            (Some(va), Some(vb)) => {
-                // Convert to JSON for comparison
-                let ja = crate::json_convert::normal_value_to_json(va).ok();
-                let jb = crate::json_convert::normal_value_to_json(vb).ok();
-
-                match (ja, jb) {
-                    (Some(JsonValue::Number(na)), Some(JsonValue::Number(nb))) => {
-                        let fa = na.as_f64().unwrap_or(0.0);
-                        let fb = nb.as_f64().unwrap_or(0.0);
-                        fa.partial_cmp(&fb).unwrap_or(Ordering::Equal)
-                    }
-                    (Some(JsonValue::String(sa)), Some(JsonValue::String(sb))) => sa.cmp(&sb),
-                    (Some(a), Some(b)) => a.to_string().cmp(&b.to_string()),
-                    _ => Ordering::Equal,
-                }
-            }
-        }
     }
 }
