+++ crates/query/src/runner/commits/grouping.rs
@@ -1,0 +1,135 @@
+use serde_json::Value as JsonValue;
+
+use crate::txn::TransactionRegistry;
+
+use super::super::{DocFetcher, QueryRunner};
+
+impl<F: DocFetcher + 'static, R: TransactionRegistry> QueryRunner<F, R> {
+    /// Generate a group key from commit field values.
+    /// Format matches Go DefraDB: `{index}_{value}_` for each field.
+    pub(super) fn generate_commit_group_key(
+        commit: &document::Document,
+        fields: &[String],
+    ) -> String {
+        let mut key = String::new();
+        for (i, field_name) in fields.iter().enumerate() {
+            key.push_str(&format!("{}_", i));
+            if let Some(value) = commit.get(field_name) {
+                if let Ok(json_val) = crate::json_convert::normal_value_to_json(value) {
+                    key.push_str(&Self::json_value_to_key(&json_val));
+                } else {
+                    key.push_str("null");
+                }
+            } else {
+                key.push_str("null");
+            }
+            key.push('_');
+        }
+        key
+    }
+
+    /// Convert a JSON value to a string for use in group key.
+    pub(super) fn json_value_to_key(value: &JsonValue) -> String {
+        match value {
+            JsonValue::Null => "null".to_string(),
+            JsonValue::Bool(b) => b.to_string(),
+            JsonValue::Number(n) => n.to_string(),
+            JsonValue::String(s) => s.clone(),
+            JsonValue::Array(arr) => {
+                format!(
+                    "[{}]",
+                    arr.iter()
+                        .map(Self::json_value_to_key)
+                        .collect::<Vec<_>>()
+                        .join(",")
+                )
+            }
+            JsonValue::Object(obj) => {
+                format!(
+                    "{{{}}}",
+                    obj.iter()
+                        .map(|(k, v)| format!("{}:{}", k, Self::json_value_to_key(v)))
+                        .collect::<Vec<_>>()
+                        .join(",")
+                )
+            }
+        }
+    }
+
+    /// Check if a JSON value matches a filter.
+    /// Used for filtering nested arrays like links and heads in commit queries.
+    pub(super) fn matches_json_filter(value: &JsonValue, filter: &crate::Filter) -> bool {
+        filter.matches_json_object(value).unwrap_or(false)
+    }
+
+    /// Build a DocumentMapping for commit fields.
+    pub(super) fn build_commits_mapping() -> crate::document::DocumentMapping {
+        let mut mapping = crate::document::DocumentMapping::new();
+        mapping.add(0, "cid");
+        mapping.add(1, "height");
+        mapping.add(2, "fieldName");
+        mapping.add(3, "docID");
+        mapping.add(4, "delta");
+        mapping.add(5, "collectionVersionId");
+        mapping.add(6, "links");
+        mapping.add(7, "heads");
+        mapping.add(8, "signature");
+        mapping
+    }
+
+    /// Convert a commit document to a fields array for filter evaluation.
+    pub(super) fn commit_to_fields(
+        commit: &document::Document,
+        _mapping: &crate::document::DocumentMapping,
+    ) -> Vec<Option<JsonValue>> {
+        let field_names = [
+            "cid",
+            "height",
+            "fieldName",
+            "docID",
+            "delta",
+            "collectionVersionId",
+            "links",
+            "heads",
+            "signature",
+        ];
+        field_names
+            .iter()
+            .map(|name| {
+                commit
+                    .get(name)
+                    .and_then(|v| crate::json_convert::normal_value_to_json(v).ok())
+            })
+            .collect()
+    }
+
+    /// Compare two JSON values for ordering.
+    pub(super) fn compare_json_values(
+        a: Option<&document::NormalValue>,
+        b: Option<&document::NormalValue>,
+    ) -> std::cmp::Ordering {
+        use std::cmp::Ordering;
+
+        match (a, b) {
+            (None, None) => Ordering::Equal,
+            (None, Some(_)) => Ordering::Less,
+            (Some(_), None) => Ordering::Greater,
+            (Some(va), Some(vb)) => {
+                // Convert to JSON for comparison
+                let ja = crate::json_convert::normal_value_to_json(va).ok();
+                let jb = crate::json_convert::normal_value_to_json(vb).ok();
+
+                match (ja, jb) {
+                    (Some(JsonValue::Number(na)), Some(JsonValue::Number(nb))) => {
+                        let fa = na.as_f64().unwrap_or(0.0);
+                        let fb = nb.as_f64().unwrap_or(0.0);
+                        fa.partial_cmp(&fb).unwrap_or(Ordering::Equal)
+                    }
+                    (Some(JsonValue::String(sa)), Some(JsonValue::String(sb))) => sa.cmp(&sb),
+                    (Some(a), Some(b)) => a.to_string().cmp(&b.to_string()),
+                    _ => Ordering::Equal,
+                }
+            }
+        }
+    }
+}
